Add files via upload
This is a modified NTP library for the Arduino platform. It uses code from several sources, including much of my own. It does not use delays. As update() is called within loop(), a packet request is made periodically, set in updatedInterval(). Once a request is made, a subsequent call to update() made after 1 second (NTP_WAIT), is made. If no response or the response is invalid, another request is made after NTP_INTERVAL.

This has been working perfectly for a couple years, but I have not tested the retry and delay logic extensively because I have not had any issues.

The second modification just completed is the retrieval and return of Kiss O Death codes. I added this because I had too many devices polling NTP servers too often, and it took some investigation to figure out what was going on. I don't have those issues now, so I have no way to test that logic because I don't want to provoke a DENY response, but for what it's worth, ChatGPT thinks it should work.